chore(RHIF-279): refactor getStatus func and edgeColumn const
